requirements.txt
bootstrap_flask==2.4.2
Flask==3.1.0
Flask_Caching==2.3.1
Flask_Login==0.6.3
flask_mail==0.10.0
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
python-dotenv==1.1.0
Requests==2.32.3
SQLAlchemy==2.0.40
Werkzeug==3.1.3
WTForms==3.2.1
